Rollback of compilation warnings to have it clean on Ubuntu.

+++ lib/src/model/outline.cpp
@@ -562,7 +562,7 @@ void Outline::addNote(Note* note, int offset)
 void Outline::addNotes(std::vector<Note*>& notesToAdd, int offset)
 {
     if(notesToAdd.size()) {
-        for(size_t i=notesToAdd.size()-1; i>=0; i--) {
+        for(int i=notesToAdd.size()-1; i>=0; i--) {
             notesToAdd[i]->makeModified();
             addNote(notesToAdd[i], offset);
         }
